refactor: extract Alibaba models, fetch, and GLM helpers into separate files
Split the monolithic provider.go and adapter_models.go into dedicated files
(models.go, models_qwen.go, models_fetch.go, models_glm.go, adapter_glm.go)
with corresponding tests. This is a pure code move — no behavioral changes.

Key behavioral fix included in this boundary: deepseek-v4-flash is no longer
classified as a reasoning model and no longer receives DashScope-specific
thinking/tool_choice overrides, matching the handling of deepseek-v4-pro.
