+++ internal/providers/alibaba/adapter_models.go
@@ -1,19 +1,13 @@
 package alibaba
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
 	"omnillm/internal/cif"
 	"omnillm/internal/providers/openaicompat"
 	"omnillm/internal/providers/shared"
 	"omnillm/internal/providers/types"
 	"strings"
-
-	"github.com/rs/zerolog/log"
 )
 
 // Adapter implements types.ProviderAdapter using openaicompat for HTTP.
@@ -71,14 +65,6 @@ func (a *Adapter) buildRequest(request *cif.CanonicalRequest, stream bool) (*ope
 			extras["enable_thinking"] = false
 		}
 	}
-	if isDeepSeekV4ModelID(model) && len(request.Tools) > 0 {
-		delete(extras, "enable_thinking")
-		extras["thinking"] = map[string]any{"type": "disabled"}
-	}
-
-	// Qwen reasoning models reject tool_choice="required" or object-style
-	// tool_choice when thinking mode is active. Strip tool_choice entirely
-	// so the upstream defaults to "auto".
 	if isQwenReasoningModel(model) && len(request.Tools) > 0 {
 		delete(extras, "enable_thinking")
 	}
@@ -101,9 +87,6 @@ func (a *Adapter) buildRequest(request *cif.CanonicalRequest, stream bool) (*ope
 	if err != nil {
 		return nil, err
 	}
-	if isDeepSeekV4ModelID(model) && len(request.Tools) > 0 {
-		chatReq.ToolChoice = nil
-	}
 	if isQwenReasoningModel(model) && len(request.Tools) > 0 {
 		chatReq.ToolChoice = nil
 	}
@@ -117,9 +100,12 @@ func (a *Adapter) buildRequest(request *cif.CanonicalRequest, stream bool) (*ope
 	if !IsReasoningModel(model) {
 		stripReasoningContent(chatReq.Messages)
 	}
+	if strings.EqualFold(RemapModel(model), "glm-5.1") {
+		normalizeGLM51Messages(chatReq)
+	}
 	// Non-reasoning third-party models require explicit empty content for
 	// tool-only assistant messages; omitting content (nil) causes a 400 error.
-	if isNonReasoningToolModel(model) && len(request.Tools) > 0 {
+	if isNonReasoningToolModel(model) {
 		ensureToolAssistantContent(chatReq.Messages)
 		ensureDashScopeToolCallAlias(chatReq.Messages)
 	}
@@ -130,117 +116,4 @@ func stripReasoningContent(messages []openaicompat.Message) {
 	for i := range messages {
 		messages[i].ReasoningContent = ""
 	}
-}
-
-func ensureToolAssistantContent(messages []openaicompat.Message) {
-	for i := range messages {
-		if messages[i].Role == "assistant" && messages[i].Content == nil && len(messages[i].ToolCalls) > 0 {
-			messages[i].Content = ""
-		}
-	}
-}
-
-func ensureDashScopeToolCallAlias(messages []openaicompat.Message) {
-	for i := range messages {
-		if messages[i].Role != "assistant" {
-			continue
-		}
-		for j := range messages[i].ToolCalls {
-			if messages[i].ToolCalls[j].ID != "" && messages[i].ToolCalls[j].CallID == "" {
-				messages[i].ToolCalls[j].CallID = messages[i].ToolCalls[j].ID
-			}
-		}
-	}
-}
-
-// ErrHardcodedFallback is returned alongside a hardcoded model list when the
-// live DashScope /models API is unavailable. Callers that cache model lists
-// should treat this as a soft error and skip caching so a future request can
-// retry the live API.
-var ErrHardcodedFallback = fmt.Errorf("alibaba: models fetch failed, using hardcoded fallback")
-
-// GetModels returns the available models for this Alibaba instance.
-// If the live API is unreachable it returns the hardcoded Qwen-only catalog
-// together with ErrHardcodedFallback so callers can decide whether to cache.
-func GetModels(instanceID, token, baseURL string, config map[string]any) (*types.ModelsResponse, error) {
-	if token == "" {
-		return nil, fmt.Errorf("alibaba: not authenticated")
-	}
-	resp, err := FetchModelsFromAPI(instanceID, token, baseURL, config)
-	if err == nil && len(resp.Data) > 0 {
-		return resp, nil
-	}
-	log.Warn().Err(err).Str("provider", instanceID).Msg("alibaba: falling back to hardcoded model list")
-	return GetModelsHardcoded(instanceID), ErrHardcodedFallback
-}
-
-// GetModelsHardcoded returns the fallback model catalog (Qwen models only).
-// DeepSeek and other third-party models available on DashScope are only
-// surfaced when FetchModelsFromAPI succeeds, because DashScope account plans
-// vary and not every key has access to every model.
-func GetModelsHardcoded(instanceID string) *types.ModelsResponse {
-	var result []types.Model
-	for _, m := range Models {
-		if strings.Contains(strings.ToLower(m.ID), "deepseek") {
-			continue // only include from live API — plan access varies
-		}
-		entry := m
-		entry.Provider = instanceID
-		result = append(result, entry)
-	}
-	return &types.ModelsResponse{Data: result, Object: "list"}
-}
-
-// FetchModelsFromAPI fetches available models from the DashScope API.
-func FetchModelsFromAPI(instanceID, token, baseURL string, _ map[string]any) (*types.ModelsResponse, error) {
-	url := strings.TrimRight(baseURL, "/") + "/models"
-	req, err := http.NewRequest(http.MethodGet, url, nil)
-	if err != nil {
-		return nil, fmt.Errorf("alibaba: failed to create models request: %w", err)
-	}
-	for k, v := range Headers(token, false, nil) {
-		req.Header.Set(k, v)
-	}
-
-	resp, err := alibabaHTTPClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("alibaba: models request failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, fmt.Errorf("alibaba: models fetch failed (%d)", resp.StatusCode)
-	}
-
-	b, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("alibaba: failed to read models response: %w", err)
-	}
-
-	var payload struct {
-		Data []struct {
-			ID string `json:"id"`
-		} `json:"data"`
-	}
-	if err := json.NewDecoder(bytes.NewReader(b)).Decode(&payload); err != nil {
-		return nil, fmt.Errorf("alibaba: failed to decode models response: %w", err)
-	}
-
-	models := make([]types.Model, 0, len(payload.Data))
-	for _, item := range payload.Data {
-		if item.ID == "" || !IsChatCompletionsModel(item.ID) {
-			continue
-		}
-		m := types.Model{ID: item.ID, Name: item.ID, Provider: instanceID}
-		if meta, ok := ModelMetadata(item.ID); ok {
-			if meta.Name != "" {
-				m.Name = meta.Name
-			}
-			m.Description = meta.Description
-			m.Capabilities = meta.Capabilities
-			m.MaxTokens = meta.MaxTokens
-		}
-		models = append(models, m)
-	}
-	return &types.ModelsResponse{Data: models, Object: "list"}, nil
 }
