+++ internal/providers/alibaba/provider.go
@@ -17,75 +17,6 @@ const UserAgent = "OmniLLM/1.0"
 // API mode constant for OpenAI-compatible DashScope endpoints.
 const AlibabaAPIModeOpenAICompatible = "openai-compatible"
 
-var Models = []types.Model{
-	{ID: "qwen3.6-max-preview", Name: "Qwen3.6 Max Preview", MaxTokens: 32768, Provider: "alibaba"},
-	{ID: "qwen3.6-plus", Name: "Qwen3.6 Plus", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3.6-flash", Name: "Qwen3.6 Flash", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3.5-plus", Name: "Qwen3.5 Plus", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3.5-omni-flash", Name: "Qwen3.5 Omni Flash", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3-coder-next", Name: "Qwen3 Coder Next", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3-coder-plus", Name: "Qwen3 Coder Plus", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3-coder-flash", Name: "Qwen3 Coder Flash", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3-max", Name: "Qwen3 Max", MaxTokens: 32768, Provider: "alibaba"},
-	{ID: "qwen3-max-preview", Name: "Qwen3 Max Preview", MaxTokens: 32768, Provider: "alibaba"},
-	{ID: "qwen3-32b", Name: "Qwen3-32B", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen3-235b-a22b-instruct", Name: "Qwen3-235B-A22B Instruct", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen-plus", Name: "Qwen Plus", MaxTokens: 131072, Provider: "alibaba"},
-	{ID: "qwen-turbo", Name: "Qwen Turbo", MaxTokens: 1000000, Provider: "alibaba"},
-	{ID: "glm-5.1", Name: "GLM 5.1", MaxTokens: 131072, Provider: "alibaba"},
-	// DeepSeek models available via DashScope — used as metadata enrichment when
-	// FetchModelsFromAPI returns them from the live /models endpoint.
-	{ID: "deepseek-v3", Name: "DeepSeek V3", MaxTokens: 65536, Provider: "alibaba"},
-	{ID: "deepseek-v4-flash", Name: "DeepSeek V4 Flash", MaxTokens: 65536, Provider: "alibaba"},
-	{ID: "deepseek-r1", Name: "DeepSeek R1", MaxTokens: 65536, Provider: "alibaba"},
-	{ID: "deepseek-r1-0528", Name: "DeepSeek R1 0528", MaxTokens: 65536, Provider: "alibaba"},
-}
-
-var reasoningModelIDs = map[string]struct{}{
-	"qwen3.6-max-preview":        {},
-	"qwen3.6-plus":               {},
-	"qwen3.6-flash":              {},
-	"qwen3-coder-next":           {},
-	"qwen3-coder-plus":           {},
-	"qwen3-coder-flash":          {},
-	"qwen3-max":                  {},
-	"qwen3-max-preview":          {},
-	"qwen3-32b":                  {},
-	"qwen3-235b-a22b-instruct":   {},
-	"qwen-plus":                  {},
-	"deepseek-r1":                {},
-	"deepseek-r1-0528":           {},
-	"deepseek-v4-flash":          {},
-}
-
-var qwenReasoningModelIDs = map[string]struct{}{
-	"qwen3.6-max-preview":        {},
-	"qwen3.6-plus":               {},
-	"qwen3.6-flash":              {},
-	"qwen3-coder-next":           {},
-	"qwen3-coder-plus":           {},
-	"qwen3-coder-flash":          {},
-	"qwen3-max":                  {},
-	"qwen3-max-preview":          {},
-	"qwen3-32b":                  {},
-	"qwen3-235b-a22b-instruct":   {},
-	"qwen-plus":                  {},
-}
-
-var deepSeekV4ModelIDs = map[string]struct{}{
-	"deepseek-v4-flash": {},
-}
-
-// nonReasoningToolModelIDs are models that are NOT reasoning models but require
-// special tool handling when tools are present on DashScope. These are typically
-// third-party models (GLM, Qwen3.5-Plus) that reject tool_choice and require
-// explicit empty content for tool-only assistant messages.
-var nonReasoningToolModelIDs = map[string]struct{}{
-	"qwen3.5-plus": {},
-	"glm-5.1":      {},
-}
-
-
 // Provider implements types.Provider for Alibaba DashScope.
 type Provider struct {
 	instanceID string
@@ -399,51 +330,4 @@ func APIKeyProviderName(config map[string]interface{}) string {
 	default:
 		return "Alibaba DashScope Standard (" + region + ")"
 	}
-}
-
-// RemapModel trims Alibaba model IDs before sending them upstream.
-func RemapModel(modelID string) string {
-	trimmed := strings.TrimSpace(modelID)
-	if idx := strings.LastIndex(trimmed, "/"); idx >= 0 {
-		return strings.TrimSpace(trimmed[idx+1:])
-	}
-	return trimmed
-}
-
-// IsChatCompletionsModel returns true if the model is not realtime-only.
-func IsChatCompletionsModel(modelID string) bool {
-	return !strings.Contains(strings.ToLower(modelID), "realtime")
-}
-
-// IsReasoningModel returns true for models that support extended thinking /
-// chain-of-thought via the enable_thinking parameter.
-func IsReasoningModel(modelID string) bool {
-	_, ok := reasoningModelIDs[strings.ToLower(RemapModel(modelID))]
-	return ok
-}
-
-func isQwenReasoningModel(modelID string) bool {
-	_, ok := qwenReasoningModelIDs[strings.ToLower(RemapModel(modelID))]
-	return ok
-}
-
-func isDeepSeekV4ModelID(modelID string) bool {
-	_, ok := deepSeekV4ModelIDs[strings.ToLower(RemapModel(modelID))]
-	return ok
-}
-
-func isNonReasoningToolModel(modelID string) bool {
-	_, ok := nonReasoningToolModelIDs[strings.ToLower(RemapModel(modelID))]
-	return ok
-}
-
-// ModelMetadata returns hardcoded metadata for a known model ID.
-func ModelMetadata(modelID string) (types.Model, bool) {
-	remapped := strings.ToLower(RemapModel(modelID))
-	for _, m := range Models {
-		if m.ID == remapped {
-			return m, true
-		}
-	}
-	return types.Model{}, false
 }
